Fixed subtle operator precedence bug in qms's phase estimations

+++ QIengine++/include/qms.cuh
@@ -275,7 +275,7 @@ void apply_phase_estimation(ComplexVec& state, const std::vector<uint>& q_state,
 
     // apply CUs
     for(int trg = q_target.size() - 1; trg > -1; --trg){
-        uint powr = (1U << (q_target.size()-1-trg));
+        double powr = (double)(1U << (q_target.size()-1-trg));
         cevolution(state, powr*t, powr*n, q_target[trg], q_state);
         suqa::apply_u1(state, q_target[trg], -powr*t*t_PE_shift);
     }
@@ -297,7 +297,7 @@ void apply_phase_estimation_inverse(ComplexVec& state, const std::vector<uint>& 
 
     // apply CUs
     for(uint trg = 0; trg < q_target.size(); ++trg){
-        uint powr = (1U << (q_target.size()-1-trg));
+        double powr = (double)(1U << (q_target.size()-1-trg));
         cevolution(state, -powr*t, powr*n, q_target[trg], q_state);
         suqa::apply_u1(state, q_target[trg], +powr*t*t_PE_shift);
     }
